# apps/api/src/cart/dto/line-item.dto.ts
import { IsInt, IsString, Min, MinLength } from "class-validator";

export class LineItemDto {
  @IsString()
  @MinLength(1)
  sku!: string;

  @IsString()
  @MinLength(1)
  name!: string;

  @IsInt()
  @Min(1)
  quantity!: number;

  @IsInt()
  @Min(0)
  unitPriceCents!: number;
}